requirements.txt
psycopg2-binary
openai
streamlit
python-dotenv
